@@ -68,42 +68,36 @@
     "aws-sdk": "^2.344.0",
     "backtrace-node": "^0.7.3",
     "crash-handler": "https://github.com/stream-labs/crash-handler/releases/download/v0.0.20/iojs-v2.0.8-crash-handler.tar.gz",
-    "electron-log": "^2.2.17",
     "electron-window-state": "^4.1.1",
     "font-manager": "https://github.com/stream-labs/font-manager/releases/download/v1.0.4/iojs-v2.0.8-font-manager.tar.gz",
     "node-fontinfo": "https://github.com/stream-labs/node-fontinfo/releases/download/v0.0.7/iojs-v2.0.8-node-fontinfo.tar.gz",
     "node-libuiohook": "https://github.com/stream-labs/node-libuiohook/releases/download/v0.0.6/iojs-v2.0.5-node-libuiohook.tar.gz",
     "obs-studio-node": "https://obsstudionodes3.streamlabs.com/obs-studio-node-v0.3.69-iojs-v2.0.8.tar.gz",
-    "recursive-readdir": "^2.2.2",
     "request": "^2.85.0",
     "rimraf": "^2.6.1",
     "semver": "^5.5.1",
     "socket.io-client": "2.1.1",
-    "uuid": "^3.0.1",
-    "vee-validate": "^2.1.0-beta.6",
-    "vue-codemirror": "^4.0.5",
-    "vue-spinner": "^1.0.3"
+    "uuid": "^3.0.1"
   },
   "devDependencies": {
     "7zip-bin": "2.4.1",
+    "@babel/core": "^7.1.6",
+    "@babel/plugin-proposal-decorators": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
     "@types/archiver": "2.1.2",
     "@types/lodash": "4.14.115",
     "@types/node": "8.10.23",
     "@types/request": "2.47.1",
     "@types/socket.io-client": "1.4.32",
-    "@types/urijs": "1.15.37",
     "@types/uuid": "3.4.3",
-    "@types/webdriverio": "4.8.6",
+    "@types/webdriverio": "^4.13.0",
     "@xkeshi/vue-qrcode": "0.3.0",
     "archiver": "2.1.1",
-    "ava": "0.25.0",
-    "babel-core": "6.26.3",
+    "ava": "v1.0.0-rc.2",
     "babel-eslint": "8.2.6",
-    "babel-loader": "7.1.5",
+    "babel-loader": "^8.0.4",
     "babel-plugin-transform-class-properties": "6.24.1",
     "babel-plugin-transform-decorators-legacy": "1.3.5",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
     "circular-dependency-plugin": "3.0.0",
     "css-element-queries": "^1.0.2",
     "css-loader": "0.28.11",
@@ -111,6 +105,7 @@
     "electron": "2.0.14",
     "electron-builder": "20.22.0",
     "electron-devtools-installer": "2.2.4",
+    "electron-log": "^2.2.17",
     "emojione": "3.1.7",
     "eslint": "3.19.0",
     "eslint-config-airbnb": "14.1.0",
@@ -131,13 +126,14 @@
     "pngjs": "3.3.3",
     "progress": "2.0.0",
     "raven-js": "3.26.4",
+    "recursive-readdir": "^2.2.2",
     "reflect-metadata": "0.1.12",
-    "rxjs": "5.5.11",
+    "rxjs": "6.3.3",
     "shx": "^0.3.2",
     "sl-vue-tree": "https://github.com/stream-labs/sl-vue-tree.git",
     "sockjs": "0.3.19",
     "sockjs-client": "1.1.5",
-    "spectron": "3.8.0",
+    "spectron": "4.0.0",
     "style-loader": "0.21.0",
     "traverse": "0.6.6",
     "ts-loader": "4.4.2",
@@ -145,11 +141,14 @@
     "tslint-config-prettier": "1.14.0",
     "tslint-loader": "3.6.0",
     "typedoc": "0.11.1",
-    "typescript": "2.9.2",
+    "typescript": "3.2.1",
+    "uglifyjs-webpack-plugin": "^2.0.1",
     "unzip-stream": "0.3.0",
-    "urijs": "1.18.8",
+    "urijs": "1.19.1",
     "v-tooltip": "2.0.0-rc.31",
+    "vee-validate": "^2.1.0-beta.6",
     "vue": "2.5.16",
+    "vue-codemirror": "^4.0.5",
     "vue-color": "2.4.3",
     "vue-i18n": "7.6.0",
     "vue-js-modal": "1.3.15",
@@ -158,12 +157,14 @@
     "vue-popperjs": "1.2.2",
     "vue-property-decorator": "7.0.0",
     "vue-slider-component": "2.2.7",
+    "vue-spinner": "^1.0.3",
     "vue-template-compiler": "2.5.16",
     "vue-toasted": "^1.1.24",
     "vuedraggable": "2.8.5",
     "vuex": "3.0.1",
-    "webpack": "4.8.3",
-    "webpack-cli": "2.1.3"
+    "webdriverio": "^4.14.1",
+    "webpack": "4.26.1",
+    "webpack-cli": "3.1.2"
   },
   "resolutions": {
     "uglify-es": "3.3.9"
