@@ -1,7 +1,7 @@
 import { IPropertyManager } from './properties-managers/properties-manager';
 import { IObsListOption, TObsFormData } from 'components/obs/inputs/ObsInput';
 import { WidgetType } from 'services/widgets';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { IAudioSource } from 'services/audio';
 
 export interface ISource extends IResource {
