@@ -2,7 +2,7 @@ import { Service } from 'services/service';
 import electron from 'electron';
 import { StreamingService, EStreamingState } from 'services/streaming';
 import { Inject } from 'util/injector';
-import { Subscription } from 'rxjs/subscription';
+import { Subscription } from 'rxjs';
 import path from 'path';
 import fs from 'fs';
 import { UsageStatisticsService } from 'services/usage-statistics';
