@@ -13,8 +13,7 @@ import {
   IScenesServiceApi
 } from './index';
 import { SourcesService, ISource } from 'services/sources';
-import electron from 'electron';
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 import { Inject } from 'util/injector';
 import * as obs from '../../../obs-api';
 import { $t } from 'services/i18n';
