@@ -14,7 +14,7 @@ import { $t } from 'services/i18n';
 import { Inject } from '../../util/injector';
 import { shortcut } from '../shortcuts';
 import { ISelection, ISelectionServiceApi, ISelectionState, TNodesList } from './selection-api';
-import { Subject } from 'rxjs/Subject';
+import { Subject } from 'rxjs';
 import Utils from '../utils';
 import { Source } from '../sources';
 import { CenteringAxis } from 'util/ScalableRectangle';
