@@ -1,12 +1,13 @@
 // Tools for dealing with forms in spectron
 
-import { GenericTestContext } from 'ava';
+import { TExecutionContext } from './index';
 
-async function getNthLabelId(t: GenericTestContext<any>, label: string, index: number) {
-  return (await t.context.app.client.$$(`label=${label}`))[index].ELEMENT;
+async function getNthLabelId(t: TExecutionContext, label: string, index: number) {
+  const el = await t.context.app.client.$$(`label=${label}`);
+  return (el[index] as any).ELEMENT;
 }
 
-export async function setFormInput(t: GenericTestContext<any>, label: string, value: string, index = 0) {
+export async function setFormInput(t: TExecutionContext, label: string, value: string, index = 0) {
   const id = await getNthLabelId(t, label, index);
 
   await t.context.app.client
@@ -14,7 +15,7 @@ export async function setFormInput(t: GenericTestContext<any>, label: string, va
     .setValue('input', value);
 }
 
-export async function getFormInput(t: GenericTestContext<any>, label: string, index = 0) {
+export async function getFormInput(t: TExecutionContext, label: string, index = 0) {
   const id = await getNthLabelId(t, label, index);
 
   return t.context.app.client
@@ -22,7 +23,7 @@ export async function getFormInput(t: GenericTestContext<any>, label: string, in
     .getValue('input');
 }
 
-export async function clickFormInput(t: GenericTestContext<any>, label: string, index = 0) {
+export async function clickFormInput(t: TExecutionContext, label: string, index = 0) {
   const id = await getNthLabelId(t, label, index);
 
   await t.context.app.client
@@ -30,7 +31,7 @@ export async function clickFormInput(t: GenericTestContext<any>, label: string, 
     .click('input');
 }
 
-export async function setFormDropdown(t: GenericTestContext<any>, label: string, value: string, index = 0) {
+export async function setFormDropdown(t: TExecutionContext, label: string, value: string, index = 0) {
   const id = await getNthLabelId(t, label, index);
 
   await t.context.app.client
@@ -43,7 +44,7 @@ export async function setFormDropdown(t: GenericTestContext<any>, label: string,
 }
 
 // Percent is a value between 0 and 1
-export async function setSliderPercent(t: GenericTestContext<any>, label: string, percent: number, index = 0) {
+export async function setSliderPercent(t: TExecutionContext, label: string, percent: number, index = 0) {
   const id = await getNthLabelId(t, label, index);
 
   const width = await t.context.app.client
@@ -53,5 +54,5 @@ export async function setSliderPercent(t: GenericTestContext<any>, label: string
 
   await t.context.app.client
     .elementIdElement(id, '../..')
-    .leftClick('.vue-slider', Math.floor(width.parsed.value * percent), 0);
+    .leftClick('.vue-slider', Math.floor(Number(width.parsed.value) * percent), 0);
 }
