@@ -1,6 +1,5 @@
 /// <reference path="../../../app/index.d.ts" />
-import 'rxjs/add/operator/first';
-import test from 'ava';
+import avaTest, { ExecutionContext, TestInterface } from 'ava';
 import { Application } from 'spectron';
 import { getClient } from '../api-client';
 import { DismissablesService } from 'services/dismissables';
@@ -10,6 +9,7 @@ const path = require('path');
 const fs = require('fs');
 const os = require('os');
 const rimraf = require('rimraf');
+export const test = avaTest as TestInterface<ITestContext>;;
 
 async function focusWindow(t: any, regex: RegExp) {
   const handles = await t.context.app.client.windowHandles();
@@ -53,13 +53,21 @@ const DEFAULT_OPTIONS: ITestRunnerOptions = {
   restartAppAfterEachTest: true
 };
 
+export interface ITestContext {
+  cacheDir: string,
+  app: Application
+}
+
+export type TExecutionContext = ExecutionContext<ITestContext>;
+
 export function useSpectron(options: ITestRunnerOptions = {}) {
   options = Object.assign({}, DEFAULT_OPTIONS, options);
   let appIsRunning = false;
   let context: any = null;
   let app: any;
+  let testPassed = false;
 
-  async function startApp(t: any) {
+  async function startApp(t: TExecutionContext) {
     t.context.cacheDir = fs.mkdtempSync(path.join(os.tmpdir(), 'slobs-test'));
     app = t.context.app = new Application({
       path: path.join(__dirname, '..', '..', '..', '..', 'node_modules', '.bin', 'electron.cmd'),
@@ -124,12 +132,18 @@ export function useSpectron(options: ITestRunnerOptions = {}) {
     appIsRunning = false;
   }
 
-  test.beforeEach(async t => {
+  test.beforeEach(async t  => {
+    testPassed = false;
     t.context.app = app;
     if (options.restartAppAfterEachTest || !appIsRunning) await startApp(t);
   });
 
+  test.afterEach(async t => {
+    testPassed = true;
+  });
+
   test.afterEach.always(async t => {
+    const testName = t.title.replace('afterEach for ', '');
     const client = await getClient();
     await client.unsubscribeAll();
     if (options.restartAppAfterEachTest) {
