@@ -1,5 +1,4 @@
-import test from 'ava';
-import { useSpectron } from '../helpers/spectron/index';
+import { test, useSpectron } from '../helpers/spectron/index';
 import { addSource } from '../helpers/spectron/sources';
 import { logIn } from '../helpers/spectron/user';
 import { FormMonkey } from '../helpers/form-monkey';
